重构删除文章

+++ app/models/article/article.go
@@ -7,12 +7,12 @@ import (
 
 // Article 文章模型
 type Article struct {
-	ID    int
+	ID    int64
 	Title string
 	Body  string
 }
 
 // Link 方法用来生成文章链接
 func (a Article) Link() string {
-	return route.Name2URL("articles.show", "id", strconv.FormatInt(int64(a.ID), 10))
+	return route.Name2URL("articles.show", "id", strconv.FormatInt(a.ID, 10))
 }
